add confirmation dialog to TopicDeleteAction

+++ src/main/java/jp/kitabatakep/intellij/plugins/codereadingrecorder/actions/TopicDeleteAction.java
@@ -4,6 +4,7 @@ import com.intellij.icons.AllIcons;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.ui.Messages;
 import jp.kitabatakep.intellij.plugins.codereadingrecorder.CodeReadingRecorderService;
 import jp.kitabatakep.intellij.plugins.codereadingrecorder.Topic;
 import org.jetbrains.annotations.NotNull;
@@ -32,8 +33,16 @@ public class TopicDeleteAction extends AnAction
     {
         CodeReadingRecorderService service = CodeReadingRecorderService.getInstance(project);
         Topic topic = topicFetcher.apply(null);
-
         if (topic == null) { return; }
-        service.getTopicList().deleteTopic(topic);
+
+        int confirmationResult = Messages.showYesNoCancelDialog(
+            "Are you sure you want to delete [" + topic.name() + "]",
+            "TopicDeleteConfirmation",
+            Messages.getQuestionIcon()
+        );
+
+        if (confirmationResult == Messages.YES) {
+            service.getTopicList().deleteTopic(topic);
+        }
     }
 }
